reviewed by: Johan Dahlin
2008-01-11  Jonathan Rogers <jonner@teegra.net>

	reviewed by: Johan Dahlin

	* src/nautilus-python-object.c:
	* src/nautilus-python.c: (nautilus_python_load_file):
	* src/nautilus-python.h:
	Use Py_ssize_t instead of int, so nautilus-python works on Python 2.5
	fixes #444366	


svn path=/trunk/; revision=48

+++ src/nautilus-python-object.c
@@ -74,7 +74,7 @@ static GObjectClass *parent_class;
 
 #define HANDLE_LIST(py_ret, type, type_name)                           \
     {                                                                  \
-        int i = 0;                                                     \
+        Py_ssize_t i = 0;                                                     \
     	if (!PySequence_Check(py_ret) || PyString_Check(py_ret)) {     \
     		PyErr_SetString(PyExc_TypeError,                           \
     						METHOD_NAME " must return a sequence");    \
